requirements.txt
pypdf==5.*
pydantic==2.*
python-dotenv==1.*
typer==0.12.*
numpy==1.26.*
scikit-learn==1.5.*
